VNSleuth v0.4: recollections (but with high storage requirements)
This is a very experimental implementation of context extension trick.
A fraction of context tokens is dedicated for "remembering" old things.
Things that would otherwise be moved out of context window and forgotten.
For determining relevancy, embeddings are computed for each line.
For computing embeddings, a special model may be specified via -md.
This release includes many other fixes.

Recommended: 500 GB of fast SSD space and 20-30 GB of free RAM for caches.
Requirements may be halved in future, and depend on the LLM used.